fix(tests): add missing list_head ISF entries and correct usize→u64 casts
- Add list_head struct+next field to deleted_exe, psaux, raw_sockets ISFs
  (walk_list() resolves list_head.next from ISF, not hardcoded offset)
- Fix tasks_offset/ctxp_offset as usize → remove cast (add_field takes u64)
  in deleted_exe, perf_event, psaux, raw_sockets
- Add missing `flags` import to kernel_timers and zombie_orphan test modules
- Commit windows/pipes.rs GUID coverage tests from Wave-2 Windows agent

+++ crates/memf-linux/src/perf_event.rs
@@ -439,8 +439,8 @@ mod tests {
         let isf = IsfBuilder::new()
             .add_symbol("init_task", sym_vaddr)
             .add_struct("task_struct", 0x400)
-            .add_field("task_struct", "tasks", tasks_offset as usize, "pointer")
-            .add_field("task_struct", "perf_event_ctxp", ctxp_offset as usize, "pointer")
+            .add_field("task_struct", "tasks", tasks_offset, "pointer")
+            .add_field("task_struct", "perf_event_ctxp", ctxp_offset, "pointer")
             .add_field("task_struct", "pid", 0x30, "unsigned int")
             .add_field("task_struct", "comm", 0x38, "char")
             .build_json();
